✂️ Welcome.js refactor: more files more fun (#2024)
