+++ frontend/components/PasteHandler.js
@@ -1,5 +1,5 @@
 import { html, useEffect } from "../imports/Preact.js"
-import { link_edit } from "./Welcome.js"
+import { link_edit } from "./welcome/Open.js"
 
 const detectNotebook = (inputtext) => {
     // Add a newline in the end for the case user didn't copy it
