+++ frontend/index.js
@@ -1,6 +1,6 @@
 import { html, render } from "./imports/Preact.js"
 import "./common/NodejsCompatibilityPolyfill.js"
 
-import { Welcome } from "./components/Welcome.js"
+import { Welcome } from "./components/welcome/Welcome.js"
 
 render(html`<${Welcome} />`, document.querySelector("main"))
